:recycle: (901): move ffmpeg audio processing into domain/audio.py
