+++ mcr-core/tests/services/speech_to_text_pipeline/test_silent_audio_detection.py
@@ -3,11 +3,11 @@ from io import BytesIO
 
 import pytest
 
-from mcr_meeting.app.exceptions.exceptions import SilentAudioError
-from mcr_meeting.app.services.audio_pre_transcription_processing_service import (
+from mcr_meeting.app.domain.audio import (
     check_audio_is_not_silent,
     compute_silence_ratio,
 )
+from mcr_meeting.app.exceptions.exceptions import SilentAudioError
 
 
 class TestComputeSilenceRatio:
